Fix Flaky unittest, allocate enough space for all needed opCodes
What does this solve?  the unit test would had failed because of creating a code buffer of size 1 to store 2 instructions.

What does it not solve: it  is still not possible to additionally set the Pc to a constant direction, this would require to generate Pc at a different point in time and account for it. This is not needed right now.