Fix Bug crashing LFVM when passing large size parameter to LOGn Operation (#541)
* Bugfix: lfvm would attempt to allocate very large memory after dynamic gas computation integer overflow.

* Handle overflow when computing mem expansion

Valid memory expansion is multiple of word size, when reaching MaxUint64 there was an overflow.  (size /32 +1 * 32)

* Add check for initially negative gas in use gas.

* Rebase + mock interface changes.

* Refactor

Reuse same words count routine in lfvm.

* Avoid overflows in UseGas routine

negative values for both existing and cost(amount), are treated as an out of gas scenario.

If this happens, it is most probably related to a previous error.