Remove use of gas on use gas failure.
The normal handling of stop conditions in the VM is to return an error and stop execution.
Errors triggered by the VM are equally critical and result of all the gas being consumed. There is no reason to consume the gas for Out fo gas scenarios but not for any of the other stopping conditions, therefore this logic is somewhere else and doing it at this point in execution is redundant and error-prone.